LayoutTests/fast/canvas/html-in-canvas/element-image-transfer-double.html

<!DOCTYPE html><!-- webkit-test-runner [ CanvasDrawElementEnabled=true ] -->
<title>TB7: postMessage with the same ElementImage twice in the transfer list throws DataCloneError</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<canvas id="cv" width="100" height="100" layoutsubtree>
  <div id="child" style="width:50px;height:50px;background:red"></div>
</canvas>
<script>
'use strict';

function waitForCanvasPaint(canvas) {
  return new Promise(resolve => {
    canvas.addEventListener('paint', resolve, {once: true});
    canvas.requestPaint();
  });
}

promise_test(async t => {
  await waitForCanvasPaint(cv);
  const ei = cv.captureElementImage(child);
  const worker = new Worker(URL.createObjectURL(new Blob([
    'self.onmessage = () => {};'
  ], {type: 'application/javascript'})));

  assert_throws_dom('DataCloneError',
    () => worker.postMessage({ image: ei }, [ei, ei]),
    'duplicate ElementImage in transfer list throws DataCloneError');

  assert_equals(ei.width, 50, 'ElementImage survives a rejected transfer (no detach)');
}, 'Duplicate ElementImage in transfer list rejects with DataCloneError; source remains usable');
</script>
